requirements.txt
torch==1.1.0          
torchvision==0.3.0
tqdm==4.32.2
tensorboard==1.14.0
Pillow==6.2.2
opencv-python==4.1.0.25
wrapt==1.11.0
scipy==1.5.1
future==0.18.2
sklearn
scikit-image
